nothing working. I'm in the middle of something
I am about to change the doorsystem list to a dictionary that only
contains an ID number and a bool. The bool to tell if the item is active
or not. There can be all kinds of items related to that list. All items
have different ids in each room